Fix throttling issues
Introduces concept of a "Frame" for batches of messages to be shown at
the same time. This ensures synchronisation between top and bottom signs
.
We also remove the watcher for termination, instead utilising the
concat() stream operator to always have the clock follow messages, but
without losing some of the final messages.
There exists a small bug whereby the first clock frame sent after a
message is not conflated correctly, resulting in the time being sent in
a separte frame to the date